+++ rasa-mysql-impl/src/test/java/eu/clarin/cmdi/rasa/TestConfig.java
@@ -99,49 +99,49 @@ public abstract class TestConfig {
       linkToBeCheckedResource = rasaFactory.getLinkToBeCheckedResource();
 
       linkToBeCheckedResource
-            .save(new LinkToBeChecked("http://www.ailla.org/waiting.html", "record", "NotGoogle", null, today));
+            .save(new LinkToBeChecked("http://www.ailla.org/waiting.html", "source", "record", "NotGoogle", null, today));
       linkToBeCheckedResource.save(
-            new LinkToBeChecked("http://www.ailla.org/audio_files/EMP1M1B1.mp3", "record", "NotGoogle", null, today));
+            new LinkToBeChecked("http://www.ailla.org/audio_files/EMP1M1B1.mp3", "source", "record", "NotGoogle", null, today));
       linkToBeCheckedResource.save(
-            new LinkToBeChecked("http://www.ailla.org/audio_files/WBA1M3A2.mp3", "record", "NotGoogle", null, today));
+            new LinkToBeChecked("http://www.ailla.org/audio_files/WBA1M3A2.mp3", "source", "record", "NotGoogle", null, today));
       linkToBeCheckedResource.save(
-            new LinkToBeChecked("http://www.ailla.org/text_files/WBA1M1A2a.mp3", "record", "NotGoogle", null, today));
+            new LinkToBeChecked("http://www.ailla.org/text_files/WBA1M1A2a.mp3", "source", "record", "NotGoogle", null, today));
       linkToBeCheckedResource.save(
-            new LinkToBeChecked("http://www.ailla.org/audio_files/KUA2M1A1.mp3", "record", "NotGoogle", null, today));
+            new LinkToBeChecked("http://www.ailla.org/audio_files/KUA2M1A1.mp3", "source", "record", "NotGoogle", null, today));
       linkToBeCheckedResource.save(
-            new LinkToBeChecked("http://www.ailla.org/text_files/KUA2M1.pdf", "record", "NotGoogle", null, today));
+            new LinkToBeChecked("http://www.ailla.org/text_files/KUA2M1.pdf", "source", "record", "NotGoogle", null, today));
       linkToBeCheckedResource.save(
-            new LinkToBeChecked("http://www.ailla.org/audio_files/sarixojani.mp3", "record", "NotGoogle", null, today));
-      linkToBeCheckedResource.save(new LinkToBeChecked("http://www.ailla.org/audio_files/TEH11M7A1sa.mp3", "record",
+            new LinkToBeChecked("http://www.ailla.org/audio_files/sarixojani.mp3", "source", "record", "NotGoogle", null, today));
+      linkToBeCheckedResource.save(new LinkToBeChecked("http://www.ailla.org/audio_files/TEH11M7A1sa.mp3", "source", "record",
             "NotGoogle", null, today));
       linkToBeCheckedResource.save(
-            new LinkToBeChecked("http://www.ailla.org/text_files/TEH11M7.pdf", "record", "NotGoogle", null, today));
+            new LinkToBeChecked("http://www.ailla.org/text_files/TEH11M7.pdf", "source", "record", "NotGoogle", null, today));
       linkToBeCheckedResource
-            .save(new LinkToBeChecked("http://dspin.dwds.de:8088/ddc-sru/dta/", "record", "NotGoogle", null, today));
+            .save(new LinkToBeChecked("http://dspin.dwds.de:8088/ddc-sru/dta/", "source", "record", "NotGoogle", null, today));
       linkToBeCheckedResource.save(
-            new LinkToBeChecked("http://dspin.dwds.de:8088/ddc-sru/grenzboten/", "record", "NotGoogle", null, today));
+            new LinkToBeChecked("http://dspin.dwds.de:8088/ddc-sru/grenzboten/", "source", "record", "NotGoogle", null, today));
       linkToBeCheckedResource
-            .save(new LinkToBeChecked("http://dspin.dwds.de:8088/ddc-sru/rem/", "record", "NotGoogle", null, today));
+            .save(new LinkToBeChecked("http://dspin.dwds.de:8088/ddc-sru/rem/", "source", "record", "NotGoogle", null, today));
       linkToBeCheckedResource.save(new LinkToBeChecked("http://www.deutschestextarchiv.de/rem/?d=M084E-N1.xml",
-            "record", "NotGoogle", null, today));
+            "source", "record", "NotGoogle", null, today));
       linkToBeCheckedResource.save(new LinkToBeChecked("http://www.deutschestextarchiv.de/rem/?d=M220P-N1.xml",
-            "record", "NotGoogle", null, today));
-      linkToBeCheckedResource.save(new LinkToBeChecked("http://www.deutschestextarchiv.de/rem/?d=M119-N1.xml", "record",
+            "source", "record", "NotGoogle", null, today));
+      linkToBeCheckedResource.save(new LinkToBeChecked("http://www.deutschestextarchiv.de/rem/?d=M119-N1.xml", "source", "record",
             "NotGoogle", null, today));
-      linkToBeCheckedResource.save(new LinkToBeChecked("http://www.deutschestextarchiv.de/rem/?d=M171-G1.xml", "record",
+      linkToBeCheckedResource.save(new LinkToBeChecked("http://www.deutschestextarchiv.de/rem/?d=M171-G1.xml", "source", "record",
             "NotGoogle", null, today));
-      linkToBeCheckedResource.save(new LinkToBeChecked("http://www.deutschestextarchiv.de/rem/?d=M185-N1.xml", "record",
+      linkToBeCheckedResource.save(new LinkToBeChecked("http://www.deutschestextarchiv.de/rem/?d=M185-N1.xml", "source", "record",
             "NotGoogle", null, today));
       linkToBeCheckedResource.save(new LinkToBeChecked("http://www.deutschestextarchiv.de/rem/?d=M048P-N1.xml",
-            "record", "NotGoogle", null, today));
-      linkToBeCheckedResource.save(new LinkToBeChecked("http://www.deutschestextarchiv.de/rem/?d=M112-G1.xml", "record",
+            "source", "record", "NotGoogle", null, today));
+      linkToBeCheckedResource.save(new LinkToBeChecked("http://www.deutschestextarchiv.de/rem/?d=M112-G1.xml", "source", "record",
             "NotGoogle", null, today));
       linkToBeCheckedResource
-            .save(new LinkToBeChecked("https://www.google.com", "GoogleRecord", "Google", null, today));
+            .save(new LinkToBeChecked("https://www.google.com", "source", "GoogleRecord", "Google", null, today));
       linkToBeCheckedResource
-            .save(new LinkToBeChecked("https://maps.google.com", "GoogleRecord", "Google", null, today));
+            .save(new LinkToBeChecked("https://maps.google.com", "source", "GoogleRecord", "Google", null, today));
       linkToBeCheckedResource
-            .save(new LinkToBeChecked("https://drive.google.com", "GoogleRecord", "Google", null, today));
+            .save(new LinkToBeChecked("https://drive.google.com", "source", "GoogleRecord", "Google", null, today));
 
       checkedLinkResource.save(new CheckedLink(null, null, "http://www.ailla.org/waiting.html", "HEAD", 200,
             "text/html; charset=UTF-8", 100l, 132, today, "Ok", 0, Category.Ok));